+++ lib/color_hue_n_tint.dart
@@ -1,7 +1,7 @@
-library color_shader;
+library color_hue_n_tint;
 
 import 'dart:core';
 import 'package:flutter/material.dart';
 
-part 'src/color_shader_base.dart';
+part 'src/color_hue_n_tint_base.dart';
 part 'src/extension/palette_list.dart';