+++ test/color_hue_n_tint_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:color_shader/color_shader.dart';
+import 'package:color_hue_n_tint/color_hue_n_tint.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('Shader Class (Default 32bits) ', () {
-    Shader shader = Shader(0xffff0000);
+  group('HueNTint Class (Default 32bits) ', () {
+    HueNTint shader = HueNTint(0xffff0000);
 
     test('.paletteList()', () {
       expect(
@@ -23,9 +23,9 @@ void main() {
           ]));
     });
 
-    test('.lightPalette()', () {
+    test('.tints()', () {
       expect(
-          shader.lightPalette(),
+          shader.tints(),
           equals([
             Color(0xffff2a2a),
             Color(0xffff5555),
@@ -35,9 +35,9 @@ void main() {
           ]));
     });
 
-    test('.darkPalette()', () {
+    test('.hues()', () {
       expect(
-          shader.darkPalette(),
+          shader.hues(),
           equals([
             Color(0xffcc0000),
             Color(0xff990000),
@@ -46,13 +46,13 @@ void main() {
           ]));
     });
 
-    test('Lightness Test', () {
-      expect(shader.lightness(scale: 0.5), equals(Color(0xffff7f7f)));
-    });
+    // test('Lightness Test', () {
+    //   expect(shader.lightness(scale: 0.5), equals(Color(0xffff7f7f)));
+    // });
 
-    test('Darkness Test', () {
-      expect(shader.darkness(scale: 0.5), equals(Color(0xff7f0000)));
-    });
+    // test('Darkness Test', () {
+    //   expect(shader.darkness(scale: 0.5), equals(Color(0xff7f0000)));
+    // });
 
     test('Default Properties Test', () {
       // print('All shades number of palette (default)');
@@ -67,8 +67,8 @@ void main() {
       expect(shader.lightShades, equals(5));
     });
   });
-  group('Shader Class (Default RGB) ', () {
-    Shader shader = Shader.fromRGB(00, 255, 169);
+  group('HueNTint Class (Default RGB) ', () {
+    HueNTint shader = HueNTint.fromRGB(00, 255, 169);
 
     test('.paletteList()', () {
       expect(
@@ -87,9 +87,9 @@ void main() {
           ]));
     });
 
-    test('.lightPalette()', () {
+    test('.tints()', () {
       expect(
-          shader.lightPalette(),
+          shader.tints(),
           equals([
             Color(0xff2affb7),
             Color(0xff55ffc5),
@@ -99,9 +99,9 @@ void main() {
           ]));
     });
 
-    test('.darkPalette()', () {
+    test('.hues()', () {
       expect(
-          shader.darkPalette(),
+          shader.hues(),
           equals([
             Color(0xff00cc87),
             Color(0xff009965),
@@ -111,8 +111,8 @@ void main() {
     });
   });
 
-  group('Shader Class (Full Scale Test)', (() {
-    Shader shader = Shader(0xffff0000, fullScale: true);
+  group('HueNTint Class (Full Scale Test)', (() {
+    HueNTint shader = HueNTint(0xffff0000, fullScale: true);
     test('.paletteList()', () {
       expect(
           shader.palette(),
@@ -129,9 +129,9 @@ void main() {
             Color(0xff000000)
           ]));
     });
-    test('.lightPalette()', () {
+    test('.tints()', () {
       expect(
-          shader.lightPalette(),
+          shader.tints(),
           equals([
             Color(0xffff3333),
             Color(0xffff6666),
@@ -140,9 +140,9 @@ void main() {
             Color(0xffffffff)
           ]));
     });
-    test('.darkPalette()', () {
+    test('.hues()', () {
       expect(
-          shader.darkPalette(),
+          shader.hues(),
           equals([
             Color(0xffbf0000),
             Color(0xff7f0000),
@@ -164,8 +164,8 @@ void main() {
       expect(shader.darkShades, equals(4));
     });
   }));
-  group('Shader Class (Full Scale RGB Test)', (() {
-    Shader shader = Shader.fromRGB(255, 0, 0, fullScale: true);
+  group('HueNTint Class (Full Scale RGB Test)', (() {
+    HueNTint shader = HueNTint.fromRGB(255, 0, 0, fullScale: true);
     test('.paletteList()', () {
       expect(
           shader.palette(),
@@ -182,9 +182,9 @@ void main() {
             Color(0xff000000)
           ]));
     });
-    test('.lightPalette()', () {
+    test('.tints()', () {
       expect(
-          shader.lightPalette(),
+          shader.tints(),
           equals([
             Color(0xffff3333),
             Color(0xffff6666),
@@ -193,9 +193,9 @@ void main() {
             Color(0xffffffff)
           ]));
     });
-    test('.darkPalette()', () {
+    test('.hues()', () {
       expect(
-          shader.darkPalette(),
+          shader.hues(),
           equals([
             Color(0xffbf0000),
             Color(0xff7f0000),
@@ -218,8 +218,8 @@ void main() {
     });
   }));
 
-  group('Shader Class (Scale value Test)', (() {
-    Shader shader = Shader(0xffff0000);
+  group('HueNTint Class (Scale value Test)', (() {
+    HueNTint shader = HueNTint(0xffff0000);
     test('.paletteList(scale: 0.5)', () {
       expect(
           shader.palette(scale: 0.5),
@@ -236,9 +236,9 @@ void main() {
             Color(0xffcc0000)
           ]));
     });
-    test('.lightPalette(scale: 0.5)', () {
+    test('.tints(scale: 0.5)', () {
       expect(
-          shader.lightPalette(scale: 0.5),
+          shader.tints(scale: 0.5),
           equals([
             Color(0xffff0a0a),
             Color(0xffff1515),
@@ -247,9 +247,9 @@ void main() {
             Color(0xffff3535)
           ]));
     });
-    test('.darkPalette(scale: 0.5)', () {
+    test('.hues(scale: 0.5)', () {
       expect(
-          shader.darkPalette(scale: 0.5),
+          shader.hues(scale: 0.5),
           equals([
             Color(0xfff20000),
             Color(0xffe50000),
@@ -259,8 +259,8 @@ void main() {
     });
   }));
 
-  group('Shader Class (Custom)', () {
-    Shader shader = Shader(0xfff2f0d9, shades: 7, index: 4);
+  group('HueNTint Class (Custom)', () {
+    HueNTint shader = HueNTint(0xfff2f0d9, shades: 7, index: 4);
     test('.paletteList()', () {
       expect(
           shader.palette(scale: 0.9),
@@ -274,9 +274,9 @@ void main() {
             Color(0xff6f6e63)
           ]));
     });
-    test('.lightPalette()', () {
+    test('.tints()', () {
       expect(
-          shader.lightPalette(scale: 0.9),
+          shader.tints(scale: 0.9),
           equals([
             Color(0xfff4f2df),
             Color(0xfff6f4e5),
@@ -284,8 +284,8 @@ void main() {
             Color(0xfffaf9f1)
           ]));
     });
-    test('.darkPalette()', () {
-      expect(shader.darkPalette(scale: 0.9),
+    test('.hues()', () {
+      expect(shader.hues(scale: 0.9),
           equals([Color(0xffb0af9e), Color(0xff6f6e63)]));
     });
 
@@ -305,7 +305,7 @@ void main() {
   });
 
   group('List Color Extension', () {
-    List<Color> palette = Shader(0xff00ff00).palette();
+    List<Color> palette = HueNTint(0xff00ff00).palette();
     test('.toListString()', () {
       expect(
           palette.toListString(),
@@ -340,26 +340,116 @@ void main() {
     });
   });
 
-  group('When Error Shader Class (Default 32bits) ', () {
-    Shader shader = Shader(0xffff0000, shades: 3, index: 5);
+  group('When Error HueNTint Class (Default 32bits) ', () {
+    HueNTint shader = HueNTint(0xffff0000, shades: 3, index: 5);
 
     test('Scale Error .palette()', () {
       expect(() => shader.palette(scale: 1.2), throwsA(isA<AssertionError>()));
       expect(() => shader.palette(scale: -1.2), throwsA(isA<AssertionError>()));
     });
 
-    test('Scale Error .lightPalette()', () {
+    test('Scale Error .tints()', () {
       expect(
-          () => shader.lightPalette(scale: 5), throwsA(isA<AssertionError>()));
+          () => shader.tints(scale: 5), throwsA(isA<AssertionError>()));
       expect(
-          () => shader.lightPalette(scale: -6), throwsA(isA<AssertionError>()));
+          () => shader.tints(scale: -6), throwsA(isA<AssertionError>()));
     });
 
-    test('Scale Error .darkPalette()', () {
+    test('Scale Error .hues()', () {
       expect(
-          () => shader.darkPalette(scale: 2), throwsA(isA<AssertionError>()));
-      expect(() => shader.darkPalette(scale: -0.5),
+          () => shader.hues(scale: 2), throwsA(isA<AssertionError>()));
+      expect(() => shader.hues(scale: -0.5),
           throwsA(isA<AssertionError>()));
     });
   });
+  group('Non full fill (Default 32bits) ', () {
+
+    test('1bit', () {
+    HueNTint shader = HueNTint(0xf, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xffaaaaaf),
+            Color(0xff55555f),
+            Color(0xff00000f),
+            Color(0xff00000a),
+            Color(0xff000005)
+          ]));
+    });
+    test('2bit', () {
+    HueNTint shader = HueNTint(0xfd, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xffaaaafe),
+            Color(0xff5555fd),
+            Color(0xff0000fd),
+            Color(0xff0000a8),
+            Color(0xff000054)
+          ]));
+    });
+    test('3bit', () {
+    HueNTint shader = HueNTint(0xfda, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xffaaaff2),
+            Color(0xff555fe6),
+            Color(0xff000fda),
+            Color(0xff000a91),
+            Color(0xff000548)
+          ]));
+    });
+    test('4bit', () {
+    HueNTint shader = HueNTint(0xfdaf, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xffaafee4),
+            Color(0xff55fdc9),
+            Color(0xff00fdaf),
+            Color(0xff00a874),
+            Color(0xff00543a)
+          ]));
+    });
+    test('5bit', () {
+    HueNTint shader = HueNTint(0xfdafd, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xffaff2fe),
+            Color(0xff5fe6fd),
+            Color(0xff0fdafd),
+            Color(0xff0a91a8),
+            Color(0xff054854)
+          ]));
+    });
+    test('6bit', () {
+    HueNTint shader = HueNTint(0xfdafdf, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xfffee4f4),
+            Color(0xfffdc9e9),
+            Color(0xfffdafdf),
+            Color(0xffa87494),
+            Color(0xff543a4a)
+          ]));
+    });
+    test('7bit', () {
+    HueNTint shader = HueNTint(0xdfdafdf, shades: 5, index: 2);
+       expect(
+          shader.palette(),
+          equals([
+            Color(0xfffee4f4),
+            Color(0xfffdc9e9),
+            Color(0xfffdafdf),
+            Color(0xffa87494),
+            Color(0xff543a4a)
+          ]));
+    });
+    });
+
+    
+
 }