+++ CHANGELOG.md
@@ -11,3 +11,14 @@
   - List\<Color>
     - .toListInt()
     - .toListString  
+
+## 1.0.1
+
+- Change name
+  - Shader Class to HueNTint Class
+    - .lightPalette() to .tints()
+    - .darkPalette() to .hues()
+    - .lightness() - delete
+    - .darkness() - delete
+- Bug fix
+